Handle talks on the scheduler table as links
Previously the link functionality (open in window or tab) of those talks was
done via javascript. Because of that they couldn't be selected via right click
to be opened in a tab, copied or bookmarked.

By embedding the talk items into a link element and adding some css to hide
the link effects, like hover and underline, we get the desired behaviour
without changing it's appearance.