Improvements in proposal controller
`if..else` is better than unless with return.
On failed update render `edit` template and not render `new` template.
On failed save redirect to index path and not render `new` template.
Change `destory` to `withdraw`, cause `destroy` makes DELETE request when
we are updating the event record.
Check for success for save in withdraw (save can fail for reasons other than validation).