Make proposal show view more robust
Partially fixes #718. Some objects have an foreign_key set on track, difficulty level or room but the
referenced object does not exists. This is caused by introducing dependent: nullify after deleting
some of the objects. This commit will make the show page more robust, however, we need to
fix the corrupt objects in an additional commit.