Support JSONP callbacks for the API/v1 endpoints
SeaGL currently uses OSEM as a backend, but still has a static website
serving as our public interface. The API endpoints have a great amount
of data, but without the jsonp callback paramater, embedding the data
is a bit difficult. This change just adds the jsonp callback param
to all the API endpoints. This should be a NOOP for anyone using direct
json calls, but will allow cross domain requests that require jsonp.