Fix issue mentioned in osem/pull/1407
After the rebase (to update the PR with the latest changes), I had the
same issue @lguerard has had. I noticed that the rails server didn't
lisen on all interfaces any more, but bound the port to the loopback
device only. By adding the parameter `-b 0.0.0.0`, I restored the
previous behavior, and OSEM could be reached with the browser after it
started up (takes a few seconds).