Remove default configuration from `.rubocop.yml`
We had part of the default Rubocop configuration in the `.rubocop.yml`
file. It is not needed as it is the default and it also don't make
sense, because we only have part of it.

That was caused because when enabling a cop we were moving it from
`.rubocop_todo.yml` to `.rubocop.yml`, but this is only needed if we
overwrite the default configuration.