adapt dir permissions for openshift and add dumb-init
This commit updates directory permissions to be compatible with
OpenShift which starts the container with an arbitrary uid which is
member of the root group. For more information please consult
https://docs.openshift.org/latest/creating_images/guidelines.html.

dumb-init is added to have a proper PID 1. bash should not run as PID 1
as this could lead to improper container shutdown.