Update mariadb to 10.2 to solve Travis failures
Our test suite is failing in Travis as it complains about not finding
`libmysqlclient-dev` which is needed for mysql2. But installing
`libmysqlclient-dev` conflicts with MariaDB 10.1. So I updated MariaDB
to 10.2 and mysql to the last version as the old one is not compatible
with MariaDB 10.2. The missed `libmysqlclient-dev` is not needed any
more.