Update yajl-ruby to 1.3.1
There is a security vulnerability in yajl-ruby 1.3.0:

In the yajl-ruby gem 1.3.0 for Ruby, when a crafted JSON file is
supplied to `Yajl::Parser.new.parse`, the whole ruby process crashes
with a `SIGABRT` in the `yajl_string_decode` function in
`yajl_encode.c`. This results in the whole ruby process terminating and
potentially a denial of service.

Reference:

https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2017-16516