Fix scheduling
Can't cache the event as the data is changed the scheduling js.

We also can't verify the token in the EventSchedules controller
as we are making susequent ajax calls and have no means to refresh
the token for the second request.

Make draggables scrollable.

There also is no need to spam the console with debug code...

Partly fixes #2353