Rephrase description of the "rspec spec" command
This might seem a bit stupid, but I got bitten by this today - I was under the
impression the command is required to make the changes visible in the running
app (despite the section above saying it isn't). Hopefully this helps any
potential contributor to get their mind sorted better than mine.