Fix flickering submit valid proposal test
The `find` was added to try to fix the flickering test:

https://github.com/openSUSE/osem/commit/ac3337bd1ed93462ee6d8cae5651fe8d633305d8

The reason why this fail is that the Javascript is not loaded yet.
`find` and `fill_in` have the same `default_max_wait_time`. Because of
that adding `find` doesn't solve the problem. Increase wait time
instead.

Fixes https://github.com/openSUSE/osem/issues/2356