Fix overlay of carousel controls
Events on the last column where not clickable because the carousel
controls where render on top. Make them narrower to solve it.

This bug was a side effect of using a carousel for rendering the
schedule. There are strong arguments to think this is not the best
design and a better solution is being discussed:

https://github.com/openSUSE/osem/issues/2421

Fixes https://github.com/openSUSE/osem/issues/2434

Co-authored-by: @sncarnera