Get rid of bin/osem-init.sh
Setting up the app if it isn't already setup should be some
concious step and not run every time we boot the containers.

This also fixes the problem that processes are not killed
if you stop the container with CTRL+C because the bash script
is not passing on signals.